requirements.txt
diff-match-patch==20121119
Django
django-appconf
django-ckeditor==5.0.3
django-debug-toolbar==1.4
django-flat-theme==1.1.3
-e git://github.com/amirrpp/django-genericadmin.git@django19#egg=genericadmin
django-grappelli==2.8.1
django-extra-views
django-image-cropping==1.0.3
django-import-export==0.4.5
django-jinja==2.1.2
django-modeladmin-reorder==0.1.3
django-mptt
django-mptt-admin
django-smart-selects==1.2.2
django-stackoverflow-trace==0.0.4
easy-thumbnails==2.3
et-xmlfile==1.0.1
jdcal==1.2
Jinja2==2.8
MarkupSafe==0.23
openpyxl
Pillow==3.2.0

PyYAML==3.11
requests
six==1.10.0
sqlparse==0.1.19
tablib==0.11.2
solid_i18n
django-crispy-forms
django-widget-tweaks
